+++ resources/client/components/SelectLanguage.vue
